refactor: 예외 처리 정보 enum class 이름 변경
