+++ backend/webtoon-core/src/main/java/com/webtoon/core/security/CustomAuthenticationEntryPoint.java
@@ -2,7 +2,7 @@ package com.webtoon.core.security;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.webtoon.core.common.ApiResponse;
-import com.webtoon.core.common.exception.ErrorType;
+import com.webtoon.core.common.exception.ExceptionType;
 import org.springframework.http.MediaType;
 import org.springframework.security.core.AuthenticationException;
 import org.springframework.security.web.AuthenticationEntryPoint;
@@ -30,13 +30,13 @@ public class CustomAuthenticationEntryPoint implements AuthenticationEntryPoint 
     @Override
     public void commence(HttpServletRequest request, HttpServletResponse response, AuthenticationException authException)
             throws IOException, ServletException {
-        ErrorType errorType = (ErrorType) request.getAttribute(AUTH_EXCEPTION);
+        ExceptionType exceptionType = (ExceptionType) request.getAttribute(AUTH_EXCEPTION);
 
         response.setContentType(MediaType.APPLICATION_JSON_VALUE);
         response.setCharacterEncoding(StandardCharsets.UTF_8.name());
-        response.setStatus(errorType.getStatus());
+        response.setStatus(exceptionType.getStatus());
 
-        String responseBody = objectMapper.writeValueAsString(ApiResponse.fail(errorType));
+        String responseBody = objectMapper.writeValueAsString(ApiResponse.fail(exceptionType));
 
         response.getWriter().print(responseBody);
     }
